leftovers from Homegirl Pro

+++ system_drive/cmd/delete.lua
@@ -1,4 +1,7 @@
 function _init(args)
+  if not args[1] then
+    return print("...delete what?")
+  end
   if not fs.delete(args[1]) then
     print("Could not delete '" .. args[1] .. "'!")
   end
